setup the base for refactoring utils and adding some performance improvements to the linter so it does not needlessly create values
