+++ src/utils/strings.ts
@@ -1,0 +1,31 @@
+/**
+ * Inserts a string at the given position in a string.
+ * @param {string} str - The string to insert into
+ * @param {number} index - The position to insert at
+ * @param {string} value - The string to insert
+ * @return {string} The string with the inserted string
+ */
+export function insert(str: string, index: number, value: string): string {
+  return str.substring(0, index) + value + str.substring(index);
+}
+
+/**
+ * Replaces a string by inserting it between the start and end positions provided for a string.
+ * @param {string} str - The string to replace a value from
+ * @param {number} start - The position to insert at
+ * @param {number} end - The position to stop text replacement at
+ * @param {string} value - The string to insert
+ * @return {string} The string with the replacement string added over the specified start and stop
+ */
+export function replaceWithValueBetweenStartAndEnd(str: string, start: number, end: number, value: string): string {
+  return str.substring(0, start) + value + str.substring(end);
+}
+
+/**
+ * Replaces \r with nothing.
+ * @param {string} text - Text to strip
+ * @return {string} Stripped text
+ */
+export function stripCr(text: string): string {
+  return text.replace(/\r/g, '');
+}
